feat: rename files in `storage` to match kebab-case and close types in /base (without those which are using outside)
